+++ src/lib.rs
@@ -53,6 +53,7 @@ pub mod scanner_pool;
 pub mod slack;
 pub mod snippet;
 pub mod sqlite;
+pub mod teams;
 pub mod update;
 pub mod util;
 pub mod validation;