+++ tests/int_gitlab.rs
@@ -115,6 +115,8 @@ fn test_gitlab_remote_scan() -> Result<()> {
             max_results: 100,
             slack_query: None,
             slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
+            teams_query: None,
+            teams_api_url: Url::parse("https://graph.microsoft.com/").unwrap(),
             // s3
             s3_bucket: None,
             s3_prefix: None,
@@ -289,6 +291,8 @@ fn test_gitlab_remote_scan_no_history() -> Result<()> {
             max_results: 100,
             slack_query: None,
             slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
+            teams_query: None,
+            teams_api_url: Url::parse("https://graph.microsoft.com/").unwrap(),
             s3_bucket: None,
             s3_prefix: None,
             role_arn: None,