+++ tests/int_validation_cache.rs
@@ -172,6 +172,8 @@ async fn test_validation_cache_and_depvars() -> Result<()> {
             max_results: 100,
             slack_query: None,
             slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
+            teams_query: None,
+            teams_api_url: Url::parse("https://graph.microsoft.com/").unwrap(),
             // s3
             s3_bucket: None,
             s3_prefix: None,