+++ tests/int_vulnerable_files.rs
@@ -115,6 +115,8 @@ impl TestContext {
                 max_results: 100,
                 slack_query: None,
                 slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
+                teams_query: None,
+                teams_api_url: Url::parse("https://graph.microsoft.com/").unwrap(),
                 // s3
                 s3_bucket: None,
                 s3_prefix: None,
@@ -259,6 +261,8 @@ impl TestContext {
                 max_results: 100,
                 slack_query: None,
                 slack_api_url: Url::parse("https://slack.com/api/").unwrap(),
+                teams_query: None,
+                teams_api_url: Url::parse("https://graph.microsoft.com/").unwrap(),
                 // s3
                 s3_bucket: None,
                 s3_prefix: None,